split models and db_models
